RLS enabled on all tables, switched cache mecahnisms
I was using unstable_cache( will soon be legacy code anyway) to cache my requests in my service layer. come to fine out enabling rls means I have to use the server client and not the browser client and cookies have to be created per request for authentication, but cookies and unstable_cache do not go together(NExtjs forbids it). This made me opt for react's 'cache', which I am hearing of for the first time by the way